resolved merge conflicts and refactor
