+++ crates/goose-mcp/src/developer/goose_hints/mod.rs
@@ -1,0 +1,2 @@
+mod import_files;
+pub mod load_hints;
